requirements.txt
google-api-python-client==2.104.0
google-auth-oauthlib==1.1.0
jupyterlab==4.0.7
python-dotenv=1.0.0
argparse=1.4.0